+++ apps/builder/app/shared/dom-hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-window-resize";
+export * from "./use-content-editable";
